refactor: consolidate SignUpPage logic into SignUp component and remove redundant file
